requirements.txt
# Server
Django>=5.2,<6
whitenoise>=6.9,<7
requests>=2.32,<3

# API Documentation
drf-spectacular>=0.28.0,<1

# Testing
djangorestframework>=3.16,<4
pytest>=8.3,<9
pytest-django>=4.11,<5
factory-boy>=3.3,<4

# Database
dj-database-url>=2.3,<3
psycopg2-binary>=2.9,<3

# Python Quality Checks
autopep8>=2.3,<3
flake8>=7.2,<8
mypy>=1.15,<2
django-stubs>=5.2,<6
djangorestframework-stubs>=3.16,<4

# Heroku
gunicorn==23.0.0